Blog_data/serializers.py
from rest_framework import serializers
from Blog_data.models import Post, Table

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'

class TableSerializer(serializers.ModelSerializer):
    class Meta:
        model = Table
        fields = '__all__'