requirements.txt
aiofiles==0.5.0
alembic==1.6.5
aniso8601==7.0.0
appdirs==1.4.4
Appium-Python-Client==1.0.2
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.5
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.3
bottle==0.12.19
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
Cython==0.29.23
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.6
django-appconf==1.0.4
django-autoslug==1.9.8
django-compressor==2.4
django-debug-toolbar==3.2
django-haystack==3.0
django-ipware==3.0.2
django-mdeditor==0.1.18
django-rest-framework==0.1.0
django-uuslug==1.2.0
djangorestframework==3.12.4
dnspython==1.16.0
dominate==2.6.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
email-validator==1.1.2
entrypoints==0.3
eventlet==0.31.1
fastapi==0.63.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Register==0.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gevent==21.1.2
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
idna==2.10
importlib-metadata==3.7.0
ipaddress==1.0.23
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.3
joblib==0.14.0
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.5.2
Mako==1.1.4
markdown2==2.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.0
opencv-python==4.5.1.48
optional-django==0.1.0
orjson==3.6.1
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
pika==1.2.0
Pillow==8.1.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.14
PyAutoGUI==0.9.52
pycparser==2.20
pydantic==1.8.1
pydicom==2.2.0
PyGetWindow==0.0.9
Pygments==2.8.0
pylint==2.7.1
pymongo==3.11.3
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynetdicom==1.5.7
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
pytesseract==0.3.8
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.0
python-editor==1.0.4
python-logstash==0.4.6
python-memcached==1.59
python-multipart==0.0.5
python-slugify==4.0.1
PyTweening==1.0.3
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
raven==6.10.0
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
Rx==1.6.1
scikit-learn==0.24.1
scipy==1.7.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.4.3
sqlparse==0.4.1
starlette==0.13.6
terminado==0.9.2
tesseract==0.1.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.3
uvicorn==0.13.4
virtualenv==20.4.1
visitor==0.1.3
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
WeRoBot==1.13.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
xgboost==0.82
xml-python==0.4.3
xmltodict==0.12.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
